add getBackgroundColor for gradient background
Implement a new function getBackgroundColor to handle the background color when a ray does not intersect with any lights or objects.
The function calculates a linear gradient from white to blue based on the vertical position of the ray, providing a visually pleasing background.
This enhances the rendering output of scenes with open sky or unobstructed views.

- This only makes the background look nicer but does not have any impact on the acctual render of the objects although this will be fixed in a later commit.
- The function may be tweeked to implement different gradients