Add a match statement in the trace_ray function for each of the possible IntersectionObject types
These changes fixes the entirety of the program but at the cost of the trace_ray function is over a 100 lines of code.

Overall these changes allows refactoring of the code and as a baseline for newer versions of the program as there is not errors in the code, if there were we could not use it as a baseline.